refactor: move type guards to modules and add examples
