+++ src/css/getCssVars.ts
@@ -3,10 +3,10 @@ import { cast } from "../internal/cast.ts";
 import { parseDOMValue } from "../internal/domValues.ts";
 import { elemOrThrow } from "../internal/elemOr.ts";
 import { formatForError } from "../internal/formatForError.ts";
-import { isCssVarName } from "../typeGuards.ts";
 import type { KeysOf, WithUndefinedValues } from "../types.ts";
 
 import { InvalidCssVarError } from "./InvalidCssVarError.ts";
+import { isCssVarName } from "./isCssVarName.ts";
 import type { CssVarName, CssVars, CssVarValue } from "./types.ts";
 
 /**
@@ -39,9 +39,15 @@ import type { CssVarName, CssVars, CssVarValue } from "./types.ts";
  * **HTML**
  *
  * ```html
- * <style>:root { --color-fg: green; }</style>
+ * <style>
+ *   :root {
+ *     --color-fg: green;
+ *   }
+ * </style>
  *
- * <button id="example" style="--color-bg: blue; --gap: 24;">Example</button>
+ * <button id="example" style="--color-bg: blue; --gap: 24;">
+ *   Example
+ * </button>
  * ```
  *
  * **Get from Element**
@@ -83,19 +89,36 @@ export function getCssVar<T extends CssVarValue>(
  * number if numeric, or the string value if a string. If not found, the value
  * is excluded from the return value.
  *
+ * **Important Note**
+ *
+ * You will need to perform checks for whether a value is `undefined` in the returned
+ * object if some of the entries weren't present.
+ *
+ * ```ts
+ * // Assuming you pass this in as the generic:
+ * type ShapeIn = {
+ *   "--color-bg": string;
+ *   "--gap": number;
+ * };
+ *
+ * // The return type of this function is:
+ * type ShapeOut = {
+ *   "--color-bg": string | undefined;
+ *   "--gap": number | undefined;
+ * };
+ * ```
+ *
  * @template T Shape of CSS variables object to return.
  *
  * @param names Names of the variable to get value for.
  * @param [target=documentElement] Optional Element, EventTarget, or CSS selector.
  *
- * @returns Object with specified names as keys and corresponding CSS variable values.
- *         Note that you will need to perform checks for whether a value is
- *          `undefined` in the returned object if some of the entries weren't present.
+ * @returns Object with specified names as keys and corresponding CSS variable values (or `undefined` if not present).
  *
  * @throws {@linkcode elems!InvalidElemError} if the specified `target` wasn't found.
  *
- * ## Examples
- * ### HTML
+ * @example
+ * **HTML**
  *
  * ```html
  * <style>
@@ -112,12 +135,12 @@ export function getCssVar<T extends CssVarValue>(
  * </button>
  * ```
  *
- * ### Get from Element
+ * **Get from Element**
  *
  * ```ts
  * type Shape = {
- *   "--color-bg": string;
- *   "--gap": number;
+ *   "--color-bg": string | undefined;
+ *   "--gap": number | undefined;
  * };
  *
  * const elem = findElem("#example")!;
@@ -126,11 +149,11 @@ export function getCssVar<T extends CssVarValue>(
  *  // { "--color-bg": "blue", "--gap": 24 }
  * ```
  *
- * ### Get from :root
+ * **Get from `:root`**
  *
  * ```ts
  * type Shape = {
- *   "--color-fg": string;
+ *   "--color-fg": string | undefined;
  * };
  *
  * getCssVars<Shape>(["--color-fg"]);
