+++ src/elems/__tests__/createElem.test.ts
@@ -56,7 +56,7 @@ describe("the createElem function", () => {
   it("creates an element with no children and properties", () => {
     const result = createElem("div", { className: "test", styles: { margin: 0, padding: 0 } });
 
-    expect(result.outerHTML).toBe(`<div class="test" style="margin: 0px; padding: 0px;"></div>`);
+    expect(result.outerHTML).toBe(`<div style="margin: 0px; padding: 0px;" class="test"></div>`);
   });
 
   it("creates an element with children and properties", () => {
