+++ src/elems/areElemsSame.ts
@@ -17,6 +17,32 @@ import type { ElemOrCssSelector } from "./types.ts";
  *             an array, returns true if **one** of the elements matches the
  *             target element input.
  * @param right Element, EventTarget, or CSS selector to compare against.
+ *
+ * @example
+ * **HTML**
+ *
+ * ```html
+ * <ul>
+ *   <li id="item-1">Item 1</li>
+ *   <li id="item-2">Item 2</li>
+ * </ul>
+ * ```
+ *
+ * **Code**
+ *
+ * ```ts
+ * const item1 = findElem("#item-1")!;
+ * const item2 = findElem("#item-2")!;
+ *
+ * areElemsSame(item1, item2);
+ * // false
+ *
+ * areElemsSame(item1, "#item-1");
+ * // true
+ *
+ * areElemsSame(item1, null);
+ * // false
+ * ```
  */
 export function areElemsSame(
   left: OneOrManyOf<ElemOrCssSelector>,
