+++ src/elems/elemExists.ts
@@ -13,6 +13,29 @@ import type { ElemOrCssSelector } from "./types.ts";
  * @param target Element, EventTarget, or CSS selector.
  *
  * @returns `true` if the specified `target` exists.
+ *
+ * @example
+ * **HTML**
+ *
+ * ```html
+ * <ul>
+ *   <li id="item-1">Item 1</li>
+ *   <li id="item-2">Item 2</li>
+ * </ul>
+ * ```
+ *
+ * **Code**
+ *
+ * ```ts
+ * elemExists("#item-1");
+ * // true
+ *
+ * elemExists(findElem("#item-1"));
+ * // true
+ *
+ * elemExists("#item-3");
+ * // false
+ * ```
  */
 export function elemExists(
   target: ElemOrCssSelector | null | undefined,
