+++ src/elems/findAllFocusable.ts
@@ -1,5 +1,6 @@
 import { findAllElems } from "./findAllElems.ts";
-import type { Elem } from "./types.ts";
+import { toElem } from "./toElem.ts";
+import type { ElemOrCssSelector } from "./types.ts";
 
 const focusableElementsSelector = [
   "a",
@@ -22,10 +23,37 @@ const focusableElementsSelector = [
  * @param [parent=document] Optional Element or EventTarget for parent.
  *
  * @returns Array of focusable elements in the optionally specified `parent`.
+ *
+ * @example
+ * **HTML**
+ *
+ * ```html
+ * <form action="" method="get" class="form-example">
+ *   <div class="form-example">
+ *     <label for="name">Enter your name: </label>
+ *     <input type="text" name="name" id="name" required />
+ *   </div>
+ *   <div class="form-example">
+ *     <label for="email">Enter your email: </label>
+ *     <input type="email" name="email" id="email" required />
+ *   </div>
+ *   <div class="form-example">
+ *     <input type="submit" value="Subscribe!" />
+ *   </div>
+ * </form>
+ * ```
+ *
+ * **Code**
+ *
+ * ```ts
+ * const items = findAllFocusable("form");
+ * // [<input id="name">, <input id="email">, <input type="submit">]
+ * ```
  */
 export function findAllFocusable(
-  parent: Elem | null | undefined = document,
+  parent: ElemOrCssSelector | null | undefined = document,
 ): HTMLElement[] {
-  // prettier-ignore
-  return findAllElems<HTMLElement>(focusableElementsSelector, parent ?? document);
+  const parentElem = toElem(parent) ?? document;
+
+  return findAllElems<HTMLElement>(focusableElementsSelector, parentElem);
 }
