+++ src/styles/getStyles.ts
@@ -1,7 +1,6 @@
 import type { AnyElement } from "../dom.ts";
 import { InvalidElemError } from "../elems/InvalidElemError.ts";
 import { toElem } from "../elems/toElem.ts";
-
 import type { ElemOrCssSelector } from "../elems/types.ts";
 import { cast } from "../internal/cast.ts";
 import { parseDOMValue } from "../internal/domValues.ts";
@@ -25,6 +24,28 @@ import type { StyleKey, Styles, StyleValue } from "./types.ts";
  *
  * @throws {@linkcode elems!InvalidElemError} if the `target` could not be found or doesn't have
  *                                            a [style](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/style) property.
+ *
+ * @example
+ * **HTML**
+ *
+ * ```html
+ * <div
+ *   id="example"
+ *   style="display: flex; line-height: 1.5;"
+ * >...</div>
+ * ```
+ *
+ * **Code**
+ *
+ * ```ts
+ * const elem = findElem("#example")!;
+ *
+ * getStyle(elem, "display");
+ * // "flex"
+ *
+ * getStyle(elem, "lineHeight");
+ * // 1.5
+ * ```
  */
 export function getStyle<T extends StyleValue>(
   target: ElemOrCssSelector,
@@ -45,17 +66,63 @@ export function getStyle<T extends StyleValue>(
  * `"false"`, a number if numeric, or the string value if a string. If not
  * found, the value is excluded from the return value.
  *
+ * **Important Note**
+ *
+ * You will need to perform checks for whether a value is `undefined` in the returned
+ * object if some of the entries weren't present.
+ *
+ * ```ts
+ * // Assuming you pass this in as the generic:
+ * type ShapeIn = {
+ *   display: string;
+ *   lineHeight: number;
+ * };
+ *
+ * // The return type of this function is:
+ * type ShapeOut = {
+ *   display: string | undefined;
+ *   lineHeight: number | undefined;
+ * };
+ * ```
+ *
  * @template T Shape of styles object to return.
  *
  * @param target Element, EventTarget, or CSS selector.
  * @param keys Names of the style properties to get values for.
  *
- * @returns Object with specified names as `keys` and corresponding style property values.
- *          Note that you will need to perform checks for whether a value is
- *          `undefined` in the returned object if some of the entries weren't present.
+ * @returns Object with specified names as `keys` and corresponding style property values (or `undefined` if not present).
  *
  * @throws {@linkcode elems!InvalidElemError} if the `target` could not be found or doesn't have
  *                                            a [style](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/style) property.
+ *
+ * @example
+ * **HTML**
+ *
+ * ```html
+ * <div
+ *   id="example"
+ *   style="display: flex; line-height: 1.5;"
+ * >...</div>
+ * ```
+ *
+ * **Code**
+ *
+ * ```ts
+ * type Shape = {
+ *   display: string | undefined;
+ *   lineHeight: number | undefined;
+ *   fontSize: number | undefined;
+ * };
+ *
+ * const elem = findElem("#example")!;
+ *
+ * getStyles<Shape>(elem, [
+ *   "display",
+ *   "lineHeight",
+ *   "fontSize",
+ * ]);
+ * // { display: "flex", lineHeight: 1.5, fontSize: undefined }
+ * ```
  */
 export function getStyles<T extends Styles = Styles>(
   target: ElemOrCssSelector,
