+++ src/styles/removeStyles.ts
@@ -1,9 +1,9 @@
+import { isCssVarName } from "../css/isCssVarName.ts";
 import type { AnyElement } from "../dom.ts";
 import type { ElemOrCssSelector } from "../elems/types.ts";
 import { cast } from "../internal/cast.ts";
 import { elemOrThrow } from "../internal/elemOr.ts";
 import { formatForError } from "../internal/formatForError.ts";
-import { isCssVarName } from "../typeGuards.ts";
 
 import type { StyleKey } from "./types.ts";
